+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpClass.cs
@@ -214,7 +214,7 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
     public CSharpClass AddProperty<TModel>(string type, TModel model, Action<CSharpProperty>? configure = null)
         where TModel : IMetadataModel, IHasName
     {
-        return AddProperty(type, model.Name, prop =>
+        return AddProperty(type, model.Name.ToCSharpIdentifier(CapitalizationBehaviour.AsIs), prop =>
         {
             prop.RepresentsModel(model);
             configure?.Invoke(prop);
@@ -232,7 +232,7 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
     public CSharpClass AddProperty<TModel>(TModel model, Action<CSharpProperty>? configure = null)
         where TModel : IMetadataModel, IHasName
     {
-        return AddProperty(File.GetModelType(model), model.Name, prop =>
+        return AddProperty(File.GetModelType(model), model.Name.ToCSharpIdentifier(CapitalizationBehaviour.AsIs), prop =>
         {
             prop.RepresentsModel(model);
             configure?.Invoke(prop);
@@ -356,7 +356,7 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
     public CSharpClass AddMethod<TModel>(string returnType, TModel model, Action<CSharpClassMethod>? configure = null)
         where TModel : IMetadataModel, IHasName
     {
-        return AddMethod(returnType, model.Name, prop =>
+        return AddMethod(returnType, model.Name.ToCSharpIdentifier(CapitalizationBehaviour.AsIs), prop =>
         {
             prop.RepresentsModel(model);
             configure?.Invoke(prop);
@@ -366,7 +366,7 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
     public CSharpClass AddMethod<TModel>(TModel returnType, Action<CSharpClassMethod>? configure = null)
         where TModel : IMetadataModel, IHasName
     {
-        return AddMethod(File.GetModelType(returnType), returnType.Name, prop =>
+        return AddMethod(File.GetModelType(returnType), returnType.Name.ToCSharpIdentifier(CapitalizationBehaviour.AsIs), prop =>
         {
             prop.RepresentsModel(returnType);
             configure?.Invoke(prop);