+++ Modules/Intent.Modules.Modelers.Domain/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 3.12.2
 
 - Improvement: Newly created entities now have a default name of *NewEntity* and not *NewClass*.
+- Improvement: Added support for `Domain` and `Services` naming conventions for `Entities`, `Attributes` and `Operations`.
 
 ### Version 3.12.1